Refactor: prisma adaptor 변경 / nextauth route.ts 파일 수정
- prisma adaptor 없어진 버전 삭제후 새로운 버전으로 변경
- build시 에러떠서 [...nextauth]/route.ts 내부의 authOption 외부로 분리
- params -> searchParams로 변경
-
