+++ package.json
@@ -12,7 +12,7 @@
         "seed": "ts-node --compiler-options {\"module\":\"CommonJS\"} prisma/seed.ts"
     },
     "dependencies": {
-        "@auth/prisma-adapter": "^1.4.0",
+        "@next-auth/prisma-adapter": "^0.4.4-canary.81",
         "@prisma/client": "^5.10.2",
         "axios": "^1.6.7",
         "next": "^14.1.3",
