+++ src/app/api/auth/[...nextauth]/auth.ts
@@ -1,0 +1,50 @@
+import prisma from "@/db/index";
+import { PrismaAdapter } from "@next-auth/prisma-adapter";
+import { NextAuthOptions } from "next-auth";
+import GoogleProvider from "next-auth/providers/google";
+import KakaoProvider from "next-auth/providers/kakao";
+import NaverProvider from "next-auth/providers/naver";
+
+const authOptions: NextAuthOptions = {
+    session: {
+        strategy: "jwt" as const,
+        maxAge: 60 * 60 * 24,
+        updateAge: 60 * 60 * 2,
+    },
+    adapter: PrismaAdapter(prisma),
+
+    providers: [
+        GoogleProvider({
+            clientId: process.env.GOOGLE_CLIENT_ID || "",
+            clientSecret: process.env.GOOGLE_CLIENT_SECRET || "",
+        }),
+        NaverProvider({
+            clientId: process.env.NAVER_CLIENT_ID || "",
+            clientSecret: process.env.NAVER_CLIENT_SECRET || "",
+        }),
+        KakaoProvider({
+            clientId: process.env.KAKAO_CLIENT_ID || "",
+            clientSecret: process.env.KAKAO_CLIENT_SECRET || "",
+        }),
+    ],
+    pages: {
+        signIn: "/users/login",
+    },
+    callbacks: {
+        session: ({ session, token }) => ({
+            ...session,
+            user: {
+                ...session.user,
+                id: token.sub,
+            },
+        }),
+        jwt: async ({ user, token }) => {
+            if (user) {
+                token.sub = user.id;
+            }
+            return token;
+        },
+    },
+};
+
+export default authOptions;
