+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,52 +1,5 @@
-import NextAuth, { NextAuthOptions } from "next-auth";
-import { PrismaAdapter } from "@auth/prisma-adapter";
-import prisma from "@/db/index";
-import GoogleProvider from "next-auth/providers/google";
-import NaverProvider from "next-auth/providers/naver";
-import KakaoProvider from "next-auth/providers/kakao";
-import { Adapter } from "next-auth/adapters";
-
-export const authOptions: NextAuthOptions = {
-    session: {
-        strategy: "jwt" as const,
-        maxAge: 60 * 60 * 24,
-        updateAge: 60 * 60 * 2,
-    },
-    adapter: PrismaAdapter(prisma) as Adapter,
-
-    providers: [
-        GoogleProvider({
-            clientId: process.env.GOOGLE_CLIENT_ID || "",
-            clientSecret: process.env.GOOGLE_CLIENT_SECRET || "",
-        }),
-        NaverProvider({
-            clientId: process.env.NAVER_CLIENT_ID || "",
-            clientSecret: process.env.NAVER_CLIENT_SECRET || "",
-        }),
-        KakaoProvider({
-            clientId: process.env.KAKAO_CLIENT_ID || "",
-            clientSecret: process.env.KAKAO_CLIENT_SECRET || "",
-        }),
-    ],
-    pages: {
-        signIn: "/users/login",
-    },
-    callbacks: {
-        session: ({ session, token }) => ({
-            ...session,
-            user: {
-                ...session.user,
-                id: token.sub,
-            },
-        }),
-        jwt: async ({ user, token }) => {
-            if (user) {
-                token.sub = user.id;
-            }
-            return token;
-        },
-    },
-};
+import NextAuth from "next-auth";
+import authOptions from "./auth";
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
