+++ src/app/page.tsx
@@ -17,9 +17,14 @@ export default async function Home() {
 }
 
 async function getData() {
-    const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/stores`, { cache: "no-store" });
-    if (!res.ok) {
-        throw new Error("Failed to fetch data");
+    try {
+        
+        const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/stores`, { cache: "no-store" });
+        if (!res.ok) {
+            throw new Error("Failed to fetch data");
+        }
+        return res.json();
+    } catch (error) {
+     console.error(error)   
     }
-    return res.json();
 }
