+++ src/pages/api/comments.ts
@@ -1,8 +1,8 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 import { getServerSession } from "next-auth";
 import prisma from "@/db";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { Comment, CommentApiResponse } from "@/interface";
+import authOptions from "@/app/api/auth/[...nextauth]/auth";
 
 interface ResponseType {
     id?: string;
