+++ src/pages/api/likes.ts
@@ -1,8 +1,8 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 import { getServerSession } from "next-auth";
 import prisma from "@/db";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { Like, LikeApiResponse } from "@/interface";
+import authOptions from "@/app/api/auth/[...nextauth]/auth";
 
 interface ResponseType {
     page?: string;
